% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/myfunc.R
\name{myfunc}
\alias{myfunc}
\title{Toy function, which raises x to the power of y}
\usage{
myfunc(x = 5, y = 2)
}
\arguments{
\item{x}{Numeric value (base). Default is 5}

\item{y}{Numeric value (exponent). Default is 2}
}
\value{
x to the power of y
}
\description{
Toy function, which raises x to the power of y
}
